enhancement(integration-tests): Move tests to new file structure
