+++ features/base/test/test_kernel_config.py
@@ -1,5 +1,7 @@
-from helper.tests.file_content import file_content
 import pytest
+from helper.sshclient import RemoteClient
+from helper.tests.file_content import file_content
+
 
 @pytest.mark.parametrize(
     "args",
@@ -23,3 +25,10 @@ def test_kernel_config(client, args, non_container):
     and 'n' are treated as the same."""
     file = "/boot/config-*"
     file_content(client, file, args, ignore_comments=True)
+
+
+def test_nvme_kernel_parameter(client, aws):
+    """ Test for NVME kernel params """
+    (exit_code, output, error) = client.execute_command("grep -c nvme_core.io_timeout=4294967295 /proc/cmdline")
+    assert exit_code == 0, f"no {error=} expected"
+    assert output.rstrip() == "1", "Expected 'nvme_core.io_timeout=4294967295' kernel parameter"
